IRGen: Factor out LegacyLayoutFormat.h and add a descriptive comment
